second-brain/backend/requirements.txt

fastapi
uvicorn
sqlalchemy
python-multipart
pydantic
qdrant-client
openai
anthropic
tiktoken
pypdf
python-docx
trafilatura
redis
arq
aiofiles
python-dotenv
aiohttp
